Fix build: remove unresolved OptIn import, use compileSdk 34, simplify CI

+++ app/src/main/java/com/musicplayer/melodex/MainActivity.kt
@@ -109,7 +109,7 @@ fun MelodexMainApp(app: MelodexApp) {
 
     // Resolve seed color
     val seedColor = remember(seedColorLong) {
-        seedColorLong?.let { Color(it) }
+        seedColorLong?.let { Color(it.toInt()) }
     }
 
     val navBackStackEntry by navController.currentBackStackEntryAsState()
